Finish footer and flex body to column
Finished up the footer to feel like I made some progress. I also
changed the div class header to hero to match the stylesheet doc
I was given. Changed the body to flex a column in aspirations of
making the vertical color bars reactive but unsure I will keep
this.